Move activity feed date utilities to local module
